Drawix_v_0/requirements.txt
fastapi
uvicorn[standard]
python-dotenv
gtts
pydub
transformers
torch
requests
python-multipart
